requirements.txt
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
diskcache==5.6.3
idna==3.6
imageio==2.33.1
lazy_loader==0.3
networkx==3.2.1
numpy==1.26.3
packaging==23.2
Pillow==9.5.0
py-mini-racer==0.6.0
PySide6==6.6.1
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.12.0
shiboken6==6.6.1
tifffile==2023.12.9
urllib3==2.1.0
